Provisional UE5 support
This accommodates changes made in the UE source to the recast library.
This hacks away at the unserialization functionality when loading
navmeshes and at least gets us loading.

A big change with UE5 is float64 usage for positional data. If we
continue with this library, we probably want to rewrite it to work with
8byte floats. I suspect right now that the navmesh would severly break
with large worls